pyproject.toml
[project]
name = "mhack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]
